feat(onboarding): account creation done
